third_party/blink/web_tests/fast/dom/custom/document-register-basic-expected.txt

CONSOLE WARNING: line 21: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Testing document.registerElement() basic behaviors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof fooConstructor is "function"
PASS fooConstructor.prototype.__proto__ is HTMLElement.prototype
PASS fooConstructor.prototype.thisIsPrototype is true
PASS document.registerElement("x-bad-a", HTMLElement) threw exception NotSupportedError: Failed to execute 'registerElement' on 'Document': Registration failed for type 'x-bad-a'. The prototype is already in-use as an interface prototype object..
PASS document.registerElement("x-bad-b", fooConstructor) threw exception NotSupportedError: Failed to execute 'registerElement' on 'Document': Registration failed for type 'x-bad-b'. The prototype is already in-use as an interface prototype object..
PASS document.registerElement("x-bad-c", { prototype: proto }) threw exception NotSupportedError: Failed to execute 'registerElement' on 'Document': Registration failed for type 'x-bad-c'. Prototype constructor property is not configurable..
PASS fooConstructor() threw exception TypeError: DOM object constructor cannot be called as a function..
PASS createdFoo.__proto__ is fooConstructor.prototype
PASS createdFoo.constructor is fooConstructor
PASS createdFoo.tagName is "X-FOO"
PASS createdFoo.textContent is "Hello"
PASS createdFoo.lastChild is childDiv
PASS parsedFoo.__proto__ is fooConstructor.prototype
PASS parsedFoo.tagName is "X-FOO"
PASS parsedFoo.someProperty is container.firstChild.someProperty
PASS barConstructor !== fooConstructor is true
PASS createdBar.tagName is "X-BAR"
PASS createdBaz.tagName is "X-BAZ"
PASS createdBaz.thisIsPrototype is true
PASS createdBaz.thisIsAlsoPrototype is true
PASS createdUpperBar.constructor is barConstructor
PASS createdUpperBar.tagName is "X-BAR"
PASS createdMixedBar.constructor is barConstructor
PASS createdMixedBar.tagName is "X-BAR"
PASS container.firstChild.constructor is barConstructor
PASS container.firstChild.tagName is "X-BAR"
PASS container.lastChild.constructor is barConstructor
PASS container.lastChild.tagName is "X-BAR"
PASS (new fooConstructor).tagName is "X-FOO"
PASS (new barConstructor).tagName is "X-BAR"
PASS (new bazConstructor).tagName is "X-BAZ"
PASS successfullyParsed is true

TEST COMPLETE
